+++ src/api/features/Images/ImageUploader.php
@@ -12,8 +12,8 @@ class ImageUploader {
 
 	public $file = [];
 	public Apikey|null $key = null;
-	public ?string $subFolder;
-	public $extensions = ["jpg", "jpeg", "png", "bmp", "webp", "wbmp"];
+	public ?string $subFolder = null;
+	public $extensions = ["jpg", "jpeg", "png", "bmp", "webp", "wbmp", "svg"];
 
 	public function SetKey(Apikey $key): ImageUploader {
 		$this->key = $key;
@@ -35,7 +35,7 @@ class ImageUploader {
 		$image = new Images();
 		$image->folder = $this->key->folder;
 		$image->upload_key = $this->key->GetID();
-		$image->subfolder = $this->subFolder;
+		if(@$this->subFolder) $image->subfolder = $this->subFolder;
 		return $image;
 	}